[SSD-121] chore: 패키지 구조 변경
