--- backend/src/main/java/com/timepaper/backend/global/auth/exception/InvalidRefreshTokenException.java
@@ -1,13 +1,0 @@
-package com.timepaper.backend.global.auth.exception;
-
-
-import com.timepaper.backend.global.exception.ErrorCode;
-import com.timepaper.backend.global.exception.custom.common.GlobalException;
-
-public class InvalidRefreshTokenException extends GlobalException {
-
-  public InvalidRefreshTokenException() {
-    super(ErrorCode.INVALID_REFRESH_TOKEN);
-  }
-
-}
