+++ backend/src/main/java/com/timepaper/backend/global/auth/service/AuthService.java
@@ -9,10 +9,10 @@ import com.timepaper.backend.global.auth.token.service.RefreshTokenService;
 import com.timepaper.backend.global.auth.token.util.JWTUtil;
 import com.timepaper.backend.global.auth.token.util.RefreshTokenUtil;
 import com.timepaper.backend.global.emailsender.EmailSendManager;
-import com.timepaper.backend.global.exception.custom.auth.DuplicateEmailException;
-import com.timepaper.backend.global.exception.custom.auth.ExpiredAuthCodeException;
-import com.timepaper.backend.global.exception.custom.auth.ExpiredEmailVerificationException;
-import com.timepaper.backend.global.exception.custom.auth.InvalidAuthCodeException;
+import com.timepaper.backend.global.exception.custom.signup.DuplicateEmailException;
+import com.timepaper.backend.global.exception.custom.signup.ExpiredAuthCodeException;
+import com.timepaper.backend.global.exception.custom.signup.ExpiredEmailVerificationException;
+import com.timepaper.backend.global.exception.custom.signup.InvalidAuthCodeException;
 import jakarta.servlet.http.HttpServletResponse;
 import java.time.Duration;
 import java.util.UUID;
