+++ backend/src/main/java/com/timepaper/backend/global/auth/token/service/RefreshTokenService.java
@@ -2,9 +2,10 @@ package com.timepaper.backend.global.auth.token.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.timepaper.backend.global.auth.exception.InvalidRefreshTokenException;
 import com.timepaper.backend.global.auth.token.entity.RefreshTokenInfo;
 import com.timepaper.backend.global.auth.token.util.RefreshTokenUtil;
+import com.timepaper.backend.global.exception.custom.auth.InvalidRefreshTokenException;
+import com.timepaper.backend.global.exception.custom.common.ServerErrorException;
 import java.time.Duration;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -37,7 +38,7 @@ public class RefreshTokenService {
     try {
       tokenInfoJson = objectMapper.writeValueAsString(tokenInfo);
     } catch (JsonProcessingException e) {
-      throw new RuntimeException("Failed to serialize RefreshTokenInfo", e);
+      throw new ServerErrorException();
     }
 
     try {
@@ -47,7 +48,7 @@ public class RefreshTokenService {
           Duration.ofDays(PERSISTENT_VALIDITY_DAYS)
       );
     } catch (DataAccessException e) {
-      throw new RuntimeException("Failed to save RefreshTokenInfo to Redis", e);
+      throw new ServerErrorException();
     }
 
   }
@@ -72,7 +73,7 @@ public class RefreshTokenService {
               SimpleGrantedAuthority::new).collect(Collectors.toList()));
 
     } catch (JsonProcessingException e) {
-      throw new RuntimeException(e);
+      throw new ServerErrorException();
     }
 
   }
