+++ backend/src/main/java/com/timepaper/backend/global/exception/custom/auth/InvalidRefreshTokenException.java
@@ -1,11 +1,13 @@
 package com.timepaper.backend.global.exception.custom.auth;
 
+
 import com.timepaper.backend.global.exception.ErrorCode;
 import com.timepaper.backend.global.exception.custom.common.GlobalException;
 
-public class DuplicateEmailException extends GlobalException {
+public class InvalidRefreshTokenException extends GlobalException {
 
-  public DuplicateEmailException() {
-    super(ErrorCode.DUPLICATE_EMAIL);
+  public InvalidRefreshTokenException() {
+    super(ErrorCode.INVALID_REFRESH_TOKEN);
   }
+
 }
