+++ backend/src/main/java/com/timepaper/backend/global/exception/custom/common/ServerErrorException.java
@@ -1,5 +1,10 @@
 package com.timepaper.backend.global.exception.custom.common;
 
-public class ServerErrorException {
+import com.timepaper.backend.global.exception.ErrorCode;
 
+public class ServerErrorException extends GlobalException {
+  
+  public ServerErrorException() {
+    super(ErrorCode.SERVER_ERROR);
+  }
 }
