+++ backend/src/main/java/com/timepaper/backend/global/exception/custom/signup/ExpiredAuthCodeException.java
@@ -1,4 +1,4 @@
-package com.timepaper.backend.global.exception.custom.auth;
+package com.timepaper.backend.global.exception.custom.signup;
 
 import com.timepaper.backend.global.exception.ErrorCode;
 import com.timepaper.backend.global.exception.custom.common.GlobalException;
